+++ app/Plugin/Schedules/Report.php
@@ -59,7 +59,28 @@ class Report extends Schedules_Base {
 	 * @return void
 	 */
 	public function run(): void {
+		// bail if setting is not enabled.
+		if ( ! $this->is_enabled() ) {
+			return;
+		}
+
+		// send the report.
 		$report_mail_object = new \PersonioIntegrationLight\Plugin\Emails\Report();
 		$report_mail_object->send();
+	}
+
+	/**
+	 * Install this schedule.
+	 *
+	 * @return void
+	 */
+	public function install(): void {
+		// bail if it is not enabled.
+		if( ! $this->is_enabled() ) {
+			return;
+		}
+
+		// install this schedule.
+		parent::install();
 	}
 }